requirements.txt
Django~=5.0
channels~=4.0.0
asgiref~=3.7.2
djangorestframework~=3.14.0
django-filter~=23.5